Handle Doubles

+++ src/Functions.vb
@@ -1,5 +1,5 @@
 ﻿Module Functions
-    Public Sub DrawPoint(X As Integer, Y As Integer, Optional SeriesRef As Integer = 0)
+    Public Sub DrawPoint(X As Double, Y As Double, Optional SeriesRef As Integer = 0)
         MainForm.Chart.Series.Item(SeriesRef).Points.AddXY(X, Y)
     End Sub
